+++ admin/js.js
@@ -2,7 +2,7 @@ jQuery(document).ready(function($) {
     /**
      * Add rating hint.
      */
-    $('body.settings_page_eml_settings h1.wp-heading-inline, body.media_page_efml_local_directories h1.wp-heading-inline').each(function() {
+    $('body.settings_page_eml_settings h1.wp-heading-inline, body.media_page_efml_local_directories h1.wp-heading-inline, body.taxonomy-edlfw_archive h1.wp-heading-inline').each(function() {
       let review_button = document.createElement('a');
       review_button.className = 'review-hint-button page-title-action';
       review_button.href = efmlJsVars.review_url;
@@ -15,7 +15,18 @@ jQuery(document).ready(function($) {
       add_file_button.href = efmlJsVars.add_file_url;
       add_file_button.innerHTML = efmlJsVars.title_add_file;
       this.after(add_file_button);
-    })
+    });
+
+    /**
+     * Add link to overview over possible external sources.
+     */
+    $('body.taxonomy-edlfw_archive h1.wp-heading-inline').each(function() {
+      let add_directory_listing_button = document.createElement( 'a' );
+      add_directory_listing_button.className = 'page-title-action';
+      add_directory_listing_button.href = efmlJsVars.directory_listing_url;
+      add_directory_listing_button.innerHTML = efmlJsVars.title_add_source;
+      this.after( add_directory_listing_button );
+    });
 
     /**
      * Add AJAX-functionality to recheck the availability of a single file.
@@ -203,6 +214,13 @@ jQuery(document).ready(function($) {
           efml_ajax_error_dialog( errorThrown )
         },
       });
+    });
+
+    /**
+     * Prevent editing of archive terms.
+     */
+    $('body.taxonomy-edlfw_archive #edittag input').each( function() {
+        $(this).attr('readonly', true);
     });
 });