+++ app/ExternalFiles/Files.php
@@ -10,7 +10,9 @@ namespace ExternalFilesInMediaLibrary\ExternalFiles;
 // prevent direct access.
 defined( 'ABSPATH' ) || exit;
 
+use easyDirectoryListingForWordPress\Directory_Listings;
 use easyDirectoryListingForWordPress\Taxonomy;
+use ExternalFilesInMediaLibrary\Plugin\Admin\Directory_Listing;
 use ExternalFilesInMediaLibrary\Plugin\Helper;
 use ExternalFilesInMediaLibrary\Plugin\Log;
 use WP_Post;
@@ -1360,71 +1362,73 @@ class Files {
 		$url = filter_input( INPUT_POST, 'url', FILTER_SANITIZE_FULL_SPECIAL_CHARS );
 
 		// bail if type or URL is not given.
-		if( is_null( $type ) || is_null( $url ) ) {
+		if ( is_null( $type ) || is_null( $url ) ) {
 			wp_send_json(
-				array( 'detail' =>
-			       array(
-						'title'     => __( 'Error', 'external-files-in-media-library' ),
-						'texts'     => array( '<p>' . __( 'The directory could not be saved as a directory archive.', 'external-files-in-media-library' ) . '</p>' ),
-						'buttons'   => array(
-							array(
-								'action'  => 'closeDialog();',
-								'variant' => 'primary',
-								'text'    => __( 'OK', 'external-files-in-media-library' ),
+				array(
+					'detail' =>
+						array(
+							'title'   => __( 'Error', 'external-files-in-media-library' ),
+							'texts'   => array( '<p>' . __( 'The directory could not be saved as a directory archive.', 'external-files-in-media-library' ) . '</p>' ),
+							'buttons' => array(
+								array(
+									'action'  => 'closeDialog();',
+									'variant' => 'primary',
+									'text'    => __( 'OK', 'external-files-in-media-library' ),
+								),
 							),
 						),
-					)
 				)
 			);
 		}
 
 		// get the login.
 		$login = filter_input( INPUT_POST, 'login', FILTER_SANITIZE_FULL_SPECIAL_CHARS );
-		if( is_null( $login ) ) {
+		if ( is_null( $login ) ) {
 			$login = '';
 		}
 
 		// get the password.
 		$password = filter_input( INPUT_POST, 'password', FILTER_SANITIZE_FULL_SPECIAL_CHARS );
-		if( is_null( $password ) ) {
+		if ( is_null( $password ) ) {
 			$password = '';
 		}
 
 		// get the API key.
 		$api_key = filter_input( INPUT_POST, 'api_key', FILTER_SANITIZE_FULL_SPECIAL_CHARS );
-		if( is_null( $api_key ) ) {
+		if ( is_null( $api_key ) ) {
 			$api_key = '';
 		}
 
 		// add the archive.
 		Taxonomy::get_instance()->add( $type, $url, $login, $password, $api_key );
 
-		// get URL for directory archive.
+		// create archive URL.
 		$url = add_query_arg(
 			array(
-				'taxonomy' => Taxonomy::get_instance()->get_name()
+				'post_type' => 'attachment',
 			),
-			get_admin_url() . 'edit-tags.php'
+			Directory_Listings::get_instance()->get_directory_archive_url()
 		);
 
 		// return OK.
 		wp_send_json(
-			array( 'detail' =>
-				array(
-					'title'     => __( 'Directory Archive saved', 'external-files-in-media-library' ),
-					'texts'     => array(
-						'<p><strong>' . __( 'The directory has been saved as archive.', 'external-files-in-media-library' ) . '</strong></p>',
-						/* translators: %1$s will be replaced by a URL. */
-						'<p>' . sprintf( __( 'You can find and use it <a href="%1$s">in the directory archive</a>.', 'external-files-in-media-library' ), $url ) . '</p>',
-					),
-					'buttons'   => array(
-						array(
-							'action'  => 'closeDialog();',
-							'variant' => 'primary',
-							'text'    => __( 'OK', 'external-files-in-media-library' ),
-						),
-					),
-				)
+			array(
+				'detail' =>
+												array(
+													'title'   => __( 'Directory Archive saved', 'external-files-in-media-library' ),
+													'texts'   => array(
+														'<p><strong>' . __( 'The directory has been saved as archive.', 'external-files-in-media-library' ) . '</strong></p>',
+														/* translators: %1$s will be replaced by a URL. */
+														'<p>' . sprintf( __( 'You can find and use it <a href="%1$s">in the directory archive</a>.', 'external-files-in-media-library' ), $url ) . '</p>',
+													),
+													'buttons' => array(
+														array(
+															'action' => 'closeDialog();',
+															'variant' => 'primary',
+															'text' => __( 'OK', 'external-files-in-media-library' ),
+														),
+													),
+												),
 			)
 		);
 	}